Extracted ASP.NET extensions into separate project so that MonoTouch can compile the core library for iPhone.
