--- src/TinyIoC/TinyIoCAspNetExtensions.cs
@@ -1,41 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Web;
-
-namespace TinyIoC
-{
-    public class HttpContextLifetimeProvider : TinyIoCContainer.ITinyIoCObjectLifetimeProvider
-    {
-        private readonly string _KeyName = String.Format("TinyIoC.HttpContext.{0}", Guid.NewGuid());
-
-        public object GetObject()
-        {
-            return HttpContext.Current.Items[_KeyName];
-        }
-
-        public void SetObject(object value)
-        {
-            HttpContext.Current.Items[_KeyName] = value;
-        }
-
-        public void ReleaseObject()
-        {
-            var item = GetObject() as IDisposable;
-
-            if (item != null)
-                item.Dispose();
-
-            SetObject(null);
-        }
-    }
-
-    public static class TinyIoCAspNetExtensions
-    {
-        public static TinyIoC.TinyIoCContainer.RegisterOptions AsPerRequestSingleton(this TinyIoC.TinyIoCContainer.RegisterOptions registerOptions)
-        {
-            return TinyIoCContainer.RegisterOptions.ToCustomLifetimeManager(registerOptions, new HttpContextLifetimeProvider(), "per request singleton");
-        }
-    }
-}
